+++ src/components/navigation/navbar/guest/Navbar.vue
@@ -1,0 +1,18 @@
+<template>
+    <base-navbar>
+        <li>
+            <a href="#" class="navbar-menu">Some Welcome Stuff</a>
+        </li>
+    </base-navbar>
+</template>
+
+<script lang="ts">
+    import BaseNavbar from "@/components/navigation/navbar/common/BaseNavbar.vue";
+
+    export default {
+        name: "LecturerNavbar",
+        components: {
+            BaseNavbar,
+        },
+    };
+</script>